requirements.txt
streamlit>=1.25
requests>=2.28
python-dateutil>=2.8.2
dateparser>=1.1.8
spacy>=3.5.0
plotly>=5.14
pandas>=2.1
openai>=1.0.0
beautifulsoup4>=4.12.2
newspaper3k>=0.2.8
tqdm>=4.65
uvicorn>=0.22.0
fastapi>=0.95.0
python-dotenv>=1.0.0
lxml
lxml_html_clean
